Move the history out of the autoupdate package
